requirements.txt
fastapi
uvicorn
typer
pydantic
python-dotenv
langchain
langchain-community
langchain-core
langchain-chroma
langchain-redis
redis
sentence-transformers
rank_bm25
pypdf
beautifulsoup4
lxml
unstructured

langchain-experimental
